requirements.txt
discord.py==2.3.0
langchain~=0.0.212
llama_index==0.6.15
nest_asyncio==1.5.6
python-dotenv~=1.0.0
BeautifulSoup4==4.12.2
soupsieve==2.4.1
asyncio~=3.4.3